Solution to day 06
First attempt at the solution I found hilarious.

It worked well for part 1 with a relatively low number of days but
in part two when faced with a high number of iterations it ate all
my RAM so I decided to let it live just for the lulz. It got to
iteration 130 or something and by then it must have had so many
ints to churn through that I decided to leave it be.

The more efficient solution for part 2 is actually amusingly simple.